+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Features:
     - SRFI 132 sorting library
 - Added SRFI 2, `and-let*`
 - Added `parameterize` from section 4.2.6 of R7RS to the `(scheme base)` library.
+- Added ` let-values` and ` let*-values` to `(scheme base)`.
 - Modified the makefile to also search current working directories for headers and libraries.
 
 Bug Fixes: